refactor: create annual review route and... ...fix issues with failing unit tests
