+++ cypress/e2e/list-management/settings-change-annual-review.feature
@@ -1,7 +1,7 @@
 Feature: List management change annual review date
 
   Background:
-    Given I am logged in as a "SuperAdmin"
+    Given I am logged in as a "Administrator"
     And A "lawyers" list exists for Eurasia
     And I click the link "Lists"
     And I click the link "Settings"
