+++ src/server/components/dashboard/controllers.ts
@@ -1,12 +1,10 @@
 import { NextFunction, Request, Response } from "express";
-import { trim, startCase } from "lodash";
-import { parseISO, format, add, isBefore, parse, isValid } from "date-fns";
+import { trim } from "lodash";
 import { dashboardRoutes } from "./routes";
 import { findUserByEmail, findUsers, isAdministrator, updateUser } from "server/models/user";
-import { createList, findListById, updateList, updateAnnualReviewDate } from "server/models/list";
+import { createList, findListById, updateList } from "server/models/list";
 import { findFeedbackByType } from "server/models/feedback";
-import { List, ServiceType, UserRoles } from "server/models/types";
-import { sendAnnualReviewDateChangeEmail } from "server/services/govuk-notify";
+
 import { isGovUKEmailAddress } from "server/utils/validation";
 import { QuestionError } from "server/components/lists";
 import { authRoutes } from "server/components/auth";
@@ -15,9 +13,11 @@ import { getCSRFToken } from "server/components/cookies/helpers";
 import { HttpException } from "server/middlewares/error-handlers";
 import { logger } from "server/services/logger";
 import { pageTitles } from "server/components/dashboard/helpers";
-import { ListIndexRes } from "server/components/dashboard/listsItems/types";
+import * as AnnualReviewHelpers from "server/components/dashboard/annualReview/helpers";
+import { UserRoles, ServiceType } from "server/models/types";
 
-const DATE_FORMAT = "d MMMM yyyy";
+import type { List } from "server/models/types";
+
 export { listItemsIndexController as listsItemsController } from "./listsItems/listItemsIndexController";
 
 export const DEFAULT_VIEW_PROPS = {
@@ -152,8 +152,11 @@ export async function listsEditController(req: Request, res: Response, next: Nex
       }
     }
 
-    const annualReviewStartDate = formatAnnualReviewDate(list as List, "annualReviewStartDate");
-    const lastAnnualReviewStartDate = formatAnnualReviewDate(list as List, "lastAnnualReviewStartDate");
+    const annualReviewStartDate = AnnualReviewHelpers.formatAnnualReviewDate(list as List, "annualReviewStartDate");
+    const lastAnnualReviewStartDate = AnnualReviewHelpers.formatAnnualReviewDate(
+      list as List,
+      "lastAnnualReviewStartDate"
+    );
 
     res.render("dashboard/lists-edit", {
       ...DEFAULT_VIEW_PROPS,
@@ -273,49 +276,6 @@ export async function listEditRemovePublisher(req: Request, res: Response): Prom
   });
 }
 
-export async function listPublisherDelete(req: Request, res: ListIndexRes, next: NextFunction): Promise<void> {
-  const userEmail = req.body.userEmail;
-  const listId = res.locals.list?.id as number;
-  const list = await findListById(listId);
-  const userHasRemovedOwnEmail = userEmail === req.user?.userData.email;
-
-  if (!list) {
-    // TODO: this should never happen. findByListId should probably throw.
-    const err = new HttpException(404, "404", "List could not be found.");
-    return next(err);
-  }
-
-  if (userHasRemovedOwnEmail) {
-    const error = {
-      field: "publisherList",
-      text: "You cannot remove your own email address from a list",
-      href: "#publishers",
-    };
-
-    return res.render("dashboard/list-edit-confirm-delete-user", {
-      ...DEFAULT_VIEW_PROPS,
-      listId: list.id,
-      userEmail,
-      error,
-      list,
-      req,
-      csrfToken: getCSRFToken(req),
-    });
-  }
-
-  const updatedUsers = list.jsonData?.users?.filter((u) => u !== userEmail) ?? [];
-
-  await updateList(list.id, { users: updatedUsers });
-
-  req.flash("changeMsg", `User ${userEmail} has been removed`);
-
-  return res.redirect(res.locals.listsEditUrl);
-}
-
-function formatAnnualReviewDate(list: List, field: string): string {
-  return list.jsonData[field] ? format(parseISO(list.jsonData[field] as string), DATE_FORMAT) : "";
-}
-
 // TODO: test
 export async function feedbackController(req: Request, res: Response, next: NextFunction): Promise<void> {
   try {
@@ -329,171 +289,4 @@ export async function feedbackController(req: Request, res: Response, next: Next
   } catch (error) {
     next(error);
   }
-}
-
-export async function listsEditAnnualReviewDateController(
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Promise<void> {
-  try {
-    const { listId } = req.params;
-    const list = await findListById(listId);
-    const annualReviewStartDate = formatAnnualReviewDate(list as List, "annualReviewStartDate");
-    const maxDate = list?.jsonData.annualReviewStartDate ? getMaxDate(list) : "";
-    const formattedMaxDate = maxDate ? format(maxDate, DATE_FORMAT) : "";
-    const helpText = maxDate
-      ? `The new date must be before ${formattedMaxDate}`
-      : "The new date must be within 6 months of today’s date";
-
-    res.render("dashboard/lists-edit-annual-review-date", {
-      ...DEFAULT_VIEW_PROPS,
-      error: {
-        text: req.flash("annualReviewError")[0],
-      },
-      annualReviewStartDate,
-      maxDate,
-      formattedMaxDate,
-      list,
-      helpText,
-      csrfToken: getCSRFToken(req),
-    });
-  } catch (error) {
-    logger.error(`listsEditAnnualReviewDateController Error: ${(error as Error).message}`);
-    next(error);
-  }
-}
-
-function getMaxDate(list: List): Date {
-  const lastAnnualReview = list.jsonData.lastAnnualReviewDate ?? list.createdAt;
-  const lastAnnualReviewPlusYear = add(lastAnnualReview as Date, { years: 1 });
-  const maxDateFromLastAnnualReview = addSixMonths(lastAnnualReviewPlusYear);
-  const maxDateFromUserEnteredValues = addSixMonths(list.jsonData.annualReviewStartDate as number);
-  const maxDate = isBefore(maxDateFromUserEnteredValues, maxDateFromLastAnnualReview)
-    ? maxDateFromUserEnteredValues
-    : maxDateFromLastAnnualReview;
-
-  return maxDate;
-}
-
-function addSixMonths(date: number | string | Date): Date {
-  const annualReviewDate = typeof date === "string" ? new Date(date) : date;
-  const newDate = add(annualReviewDate as Date, { months: 6 });
-
-  return newDate;
-}
-
-export async function listsEditAnnualReviewDatePostController(
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Promise<void> {
-  try {
-    return req.body.action === "confirmNewDate"
-      ? await confirmNewAnnualReviewDate(req, res)
-      : await updateNewAnnualReviewDate(req, res);
-  } catch (error) {
-    logger.error(`listsEditAnnualReviewDatePostController Error: ${(error as Error).message}`);
-    next(error);
-  }
-}
-
-async function confirmNewAnnualReviewDate(req: Request, res: Response): Promise<void> {
-  const { listId } = req.params;
-  const list = (await findListById(listId)) as List;
-  const { day, month } = req.body;
-  const annualReviewDate = getAnnualReviewDate({ day, month, list });
-
-  if (!annualReviewDate.value) {
-    req.flash("annualReviewError", annualReviewDate.errorMsg!);
-    return res.redirect(`${dashboardRoutes.listsEditAnnualReviewDate.replace(":listId", list.id.toString())}`);
-  }
-
-  return res.render("dashboard/lists-edit-annual-review-date-confirm", {
-    ...DEFAULT_VIEW_PROPS,
-    newAnnualReviewDateFormatted: format(annualReviewDate.value, DATE_FORMAT),
-    newAnnualReviewDate: annualReviewDate.value,
-    list,
-    csrfToken: getCSRFToken(req),
-  });
-}
-
-export function getAnnualReviewDate({ day, month, list }: { day: string; month: string; list: List }): {
-  value: Date | null;
-  errorMsg: string | null;
-} {
-  const lastAnnualReview = list.jsonData.lastAnnualReviewDate ?? list.createdAt;
-
-  const annualReviewYear = getAnnualReviewYear({
-    day,
-    month,
-    lastAnnualReview: (list.jsonData.annualReviewStartDate ?? lastAnnualReview) as number,
-  });
-  const parsedDate = parse(`${month}/${day}/${annualReviewYear}`, "P", new Date());
-
-  const maxDate = getMaxDate(list);
-
-  const invalidResult = { value: null };
-  const isLeapYear = month === "2" && day === "29";
-  let errorMsg = null;
-
-  if (!maxDate) throw new Error("confirmNewAnnualReviewDate Error: Max date could not be calculated");
-
-  if (!month || !day) {
-    errorMsg = "Enter a date for the annual review";
-    return { ...invalidResult, errorMsg };
-  }
-
-  if (isLeapYear || !isValid(parsedDate)) {
-    errorMsg = "You cannot set the annual review to this date. Please choose another";
-    return { ...invalidResult, errorMsg };
-  }
-
-  if (!isBefore(parsedDate, maxDate)) {
-    errorMsg = "You can only change the date up to 6 months after the current review date";
-    return { ...invalidResult, errorMsg };
-  }
-
-  return { value: parsedDate, errorMsg };
-}
-
-function getAnnualReviewYear({
-  day,
-  month,
-  lastAnnualReview,
-}: {
-  day?: string;
-  month?: string;
-  lastAnnualReview: number;
-}): number {
-  const date = new Date(lastAnnualReview);
-
-  if (!day || !month) return date.getFullYear();
-
-  const userEnteredDate = new Date(`${month}/${day}/${date.getFullYear()}`);
-
-  return isBefore(userEnteredDate, date) ? date.getFullYear() + 1 : date.getFullYear();
-}
-
-async function updateNewAnnualReviewDate(req: Request, res: Response): Promise<void> {
-  const { listId } = req.params;
-  const list = (await findListById(listId)) as List;
-  const { newAnnualReviewDate } = req.body;
-  const newAnnualReviewDateFormatted = new Date(newAnnualReviewDate as string);
-  const annualReviewDate = format(newAnnualReviewDateFormatted, DATE_FORMAT);
-
-  await updateAnnualReviewDate(listId, newAnnualReviewDateFormatted.toISOString());
-
-  for (const emailAddress of list.jsonData.users ?? []) {
-    await sendAnnualReviewDateChangeEmail({
-      emailAddress,
-      serviceType: startCase(list.type),
-      country: list.country!.name!,
-      annualReviewDate,
-    });
-  }
-
-  return res.redirect(
-    `${dashboardRoutes.listsEdit.replace(":listId", list.id.toString())}?annualReviewDateUpdated=true`
-  );
 }
