+++ src/server/models/types.ts
@@ -29,6 +29,12 @@ export interface ListJsonData extends JsonObject {
   lastAnnualReviewStartDate?: string;
 }
 export interface List extends PrismaClient.List {
+  id: number;
+  reference: string;
+  createdAt: Date;
+  updatedAt: Date;
+  type: string;
+  countryId: number;
   jsonData: ListJsonData;
   country?: Partial<Country>;
 }
