+++ src/server/views/dashboard/lists.njk
@@ -37,9 +37,7 @@
             <td class="govuk-table__cell govuk-table__cell--numeric"></td>
 
             <td class="govuk-table__cell govuk-table__cell--numeric">
-              {% if userIsListPublisher(req, list) %}
                 <a href="{{ dashboardRoutes.listsEdit.replace(':listId', list.id) }}" class="govuk-link">Settings</a>
-              {% endif %}
             </td>
           </tr>
         {% endfor %}
